backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
easyocr
pillow
numpy
python-dotenv
